[vulkan] indirect draw, nonuniformEXT, multi-instanced rendering

+++ Nork/Source/Main/VulkanContext.h
@@ -32,6 +32,7 @@ public:
         {
             extensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
         }
+        //extensions.push_back(VK_KHR_GET_PHYSICAL_DEVICE_PROPERTIES_2_EXTENSION_NAME);
 
         return extensions;
     }
@@ -80,7 +81,7 @@ public:
         createInfo.pApplicationInfo = &appInfo;
 
         auto extensions = getRequiredExtensions();
-        createInfo.enabledExtensionCount = static_cast<uint32_t>(extensions.size());
+        createInfo.enabledExtensionCount = extensions.size();
         createInfo.ppEnabledExtensionNames = extensions.data();
 
         if (enableValidationLayers && !checkValidationLayerSupport())
